requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.0
PyMuPDF==1.24.14
sqlparse==0.5.2
tzdata==2024.2
